Changed default int literal type from i32 -> i64, and changed ecx version of low 2 bytes from CL to CX in the newly added derefnode-case in GenerateFunctionBody.

+++ BongusCode_Compiler/src/AST/ASTNode.h
@@ -83,7 +83,7 @@ namespace AST
 		inline const ui64 Get(void) const { return n; }
 		friend Node* MakeIntNode(i32);
 
-		static const PrimitiveType s_defaultIntLiteralType = PrimitiveType::i32;
+		static const PrimitiveType s_defaultIntLiteralType = PrimitiveType::i64;
 
 	private:
